Add preliminary preview functionality
This commit adds my first attempt at a preview feature to this script.
The reason this is useful is that if the selected connector is a part of
a domain that takes some time to re-initialize, it can take a while for
the Apply button to render the requested change, and, if the user wants
to tweak the applied spacing, he/she must wait for the domain to be
re-initialized to see every subsequent change.

This commit adds a Preview button which allows the user to see
(temporarily for now) the requested change to the connector without
having to wait for the domain to re-initialize.

This commit is not how I want this to work, but it is a step in the
correct direction.